pyproject.toml
[tool.poetry]
name = "outris"
version = "0.1.0"
description = "Natural language CLI for any API - register, query, and collaborate"
authors = ["Outris Team <hello@outris.dev>"]
readme = "README.md"
homepage = "https://outris.dev"
repository = "https://github.com/outris/cli"
keywords = ["api", "cli", "openapi", "natural-language", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.9"
typer = "0.7.0"
rich = "^13.0.0"
requests = "^2.31.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
keyring = "^24.0.0"
click = "8.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
mypy = "^1.5.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
outris = "outris.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
